# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Swagger warehouse API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:1323",
    "paths": {
...
        "/stockmovementvehicle/retail": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stockmovementvehicle"
                ],
                "parameters": [
                    {
                        "description": "json req body",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateStockmovementvehicleRetail"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/stockmovementvehicle/retail/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stockmovementvehicle"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "json req body",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateStockmovementvehicleRetail"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.CreateStockmovementvehicleRetail": {
            "type": "object",
            "required": [
                "fromWarehouseId",
                "productId",
                "retailId"
            ],
            "properties": {
                "driverName": {
                    "type": "string"
                },
                "fromWarehouseId": {
                    "type": "string"
                },
                "isNewVehiclerdriver": {
                    "type": "boolean"
                },
                "nik": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "plateNumber": {
                    "type": "string"
                },
                "productId": {
                    "type": "string"
                },
                "retailId": {
                    "type": "string"
                },
                "sentGrossQuantity": {
                    "type": "number"
                },
                "sentNetQuantity": {
                    "type": "number"
                },
                "sentTareQuantity": {
                    "type": "number"
                },
                "vehicleId": {
                    "type": "string"
                },
                "vehicleName": {
                    "type": "string"
                }
            }
        },
...
        "request.UpdateStockmovementvehicleRetail": {
            "type": "object",
            "properties": {
                "sentGrossQuantity": {
                    "type": "number"
                },
                "sentNetQuantity": {
                    "type": "number"
                },
                "sentTareQuantity": {
                    "type": "number"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "payload": {
                    "type": "object"
                },
                "status": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}